added customize html template
